Moved typ declarations out to worker.go
