+++ pkg/worker/builder.go
@@ -6,51 +6,7 @@ import (
 	"time"
 
 	"github.com/iver-wharf/wharf-cmd/pkg/core/wharfyml"
-	"github.com/iver-wharf/wharf-core/pkg/logger"
 )
-
-// TODO: Change to factory pattern so the steps can get validated before
-// it's their turn
-
-var log = logger.New()
-
-type BuildOptions struct {
-	StageFilter string
-}
-
-type Builder interface {
-	Build(ctx context.Context, def wharfyml.BuildDefinition, opt BuildOptions) (Result, error)
-}
-
-type StageRunner interface {
-	RunStage(ctx context.Context, def wharfyml.Stage) StageResult
-}
-
-type StepRunner interface {
-	RunStep(ctx context.Context, def wharfyml.Step) StepResult
-}
-
-type Result struct {
-	Status   Status
-	Options  BuildOptions
-	Stages   []StageResult
-	Duration time.Duration
-}
-
-type StageResult struct {
-	Name     string
-	Status   Status
-	Steps    []StepResult
-	Duration time.Duration
-}
-
-type StepResult struct {
-	Name     string
-	Status   Status
-	Type     string
-	Error    error
-	Duration time.Duration
-}
 
 type builder struct {
 	stageRun StageRunner
@@ -97,6 +53,7 @@ func (b builder) Build(ctx context.Context, def wharfyml.BuildDefinition, opt Bu
 				WithStringf("stages", "%d/%d", stagesDone, stagesCount).
 				WithString("stage", res.Name).
 				WithDuration("dur", res.Duration.Truncate(time.Second)).
+				WithStringer("status", res.Status).
 				WithString("failed", strings.Join(failed, ",")).
 				WithString("cancelled", strings.Join(cancelled, ",")).
 				Message("Failed stage. Skipping any further stages.")
